# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema PetShop",
        "version": "2.02.3"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
        "/especie/select": {
            "get": {
                "tags": [
                    "Especie"
                ],
                "operationId": "bd36b19766ba49a5226ff9a1ae479b61",
                "responses": {
                    "200": {
                        "description": "Apreseta todas as Espécies"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/medicamento/cadastro": {
            "post": {
                "tags": [
                    "Medicamento"
                ],
                "operationId": "e183d8d9589ff03f389f9a737fb9aa67",
                "parameters": [
                    {
                        "name": "descricao",
                        "required": "true"
                    },
                    {
                        "name": "tipo_medicamento",
                        "required": "true"
                    },
                    {
                        "name": "valor",
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cadastra os Medicamentos"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/medicamento/listar/": {
            "get": {
                "tags": [
                    "Medicamento"
                ],
                "operationId": "3c8ae93bdbea64018e36462d58870a94",
                "responses": {
                    "200": {
                        "description": "Lista os medicamentos cadastrados"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/medicamento/atualizar/{id}": {
            "put": {
                "tags": [
                    "Medicamento"
                ],
                "operationId": "7fc5c2bebeb79b290f485e97b4da075f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "/atualizar/{id}",
                        "description": "Medicameto id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "descricao",
                        "required": "true"
                    },
                    {
                        "name": "tipo_medicamento",
                        "required": "true"
                    },
                    {
                        "name": "valor",
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Atualiza o Medicamento"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}